feat: support systemConfigs.${currentSystem}.${attr}
This implements support for multiarch systemConfigs in the style of
flake-utils &c.

First, it checks for `systemConfigs.${currentSystem}.${attr}`.
`currentSystem` is evaluated ahead of time in a separate Nix invocation
for simplicitly. Then, if that does not exist, it falls back to
`systemConfigs.${attr}`.

The order of evaluation is like this primarily to avoid issues where,
say, `attr == "x86_64-linux"`.

Closes #230