feat: Add init subcommand (#210)
Closes #205

Adds a subcommand for initializing a new configuration at a given
path, similarly to home-manager initializing `home.nix` at
`~/.config/home-manager`. The template configuration files will
be available under the `templates` attribute on the flake, and included
in the system-manager binary to avoid unnecessary network calls.

Thanks to @mike-lloyd03 for their contribution which inspired this change.